+++ dgisim/src/character/character.py
@@ -48,6 +48,7 @@ __all__ = [
     "Venti",
     "Xingqiu",
     "YaeMiko",
+    "Yoimiya",
 ]
 
 
@@ -2076,6 +2077,88 @@ class YaeMiko(Character):
             max_hp=10,
             energy=0,
             max_energy=2,
+            hiddens=stts.Statuses(()),
+            equipments=stts.EquipmentStatuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
+
+class Yoimiya(Character):
+    _ELEMENT = Element.PYRO
+    _WEAPON_TYPE = WeaponType.BOW
+    _TALENT_STATUS = stt.NaganoharaMeteorSwarmStatus
+    _FACTIONS = frozenset((Faction.INAZUMA,))
+
+    _NORMAL_ATTACK_COST = AbstractDices({
+        Element.PYRO: 1,
+        Element.ANY: 2,
+    })
+    _ELEMENTAL_SKILL1_COST = AbstractDices({
+        Element.PYRO: 1,
+    })
+    _ELEMENTAL_BURST_COST = AbstractDices({
+        Element.PYRO: 3,
+    })
+
+    @override
+    def _normal_attack(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    @override
+    def _elemental_skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.AddCharacterStatusEffect(
+                target=source,
+                status=stt.NiwabiEnshouStatus,
+            ),
+        )
+
+    @override
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.get_max_energy(),
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.PYRO,
+                damage=3,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+            eft.AddCombatStatusEffect(
+                target_pid=source.pid,
+                status=stt.AurousBlazeStatus,
+            ),
+        )
+
+    @override
+    def _post_elemental_skill1(
+            self,
+            game_state: GameState,
+            source: StaticTarget,
+            effects: tuple[eft.Effect, ...]
+    ) -> tuple[eft.Effect, ...]:
+        """ Removes energy recharge effect """
+        return effects + (
+            eft.AliveMarkCheckerEffect(),
+        )
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=3,
             hiddens=stts.Statuses(()),
             equipments=stts.EquipmentStatuses(()),
             statuses=stts.Statuses(()),