+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 
 - New Characters:
   - Fatui Pyro Agent
+  - Yoimiya
 - New Cards:
   - Equipment Card:
     - Talent Equipment Card:
+      - Naganohara Meteor Swarm
       - Paid in Full
 
 ## v0.3.1 (28 Aug 2023)